+++ src/BlurHash.php
@@ -10,118 +10,77 @@
 
 namespace dodecastudio\blurhash;
 
+use dodecastudio\blurhash\models\Settings;
+use dodecastudio\blurhash\gql\directives\BlurHashToUri;
+use dodecastudio\blurhash\gql\directives\AssetToBlurHash;
+use dodecastudio\blurhash\services\BlurHashService;
 use dodecastudio\blurhash\twigextensions\BlurHashTwigExtension;
 
 use Craft;
 use craft\base\Plugin;
+use craft\services\Gql;
 use craft\services\Plugins;
-use craft\events\PluginEvent;
-
+use craft\events\RegisterGqlDirectivesEvent;
 use yii\base\Event;
 
+
 /**
- * Craft plugins are very much like little applications in and of themselves. We’ve made
- * it as simple as we can, but the training wheels are off. A little prior knowledge is
- * going to be required to write a plugin.
- *
- * For the purposes of the plugin docs, we’re going to assume that you know PHP and SQL,
- * as well as some semi-advanced concepts like object-oriented programming and PHP namespaces.
- *
- * https://docs.craftcms.com/v3/extend/
- *
+ * 
  * @author    Dodeca Studio
  * @package   BlurHash
- * @since     1.0.0
+ * @since     1.0.1
  *
  */
 class BlurHash extends Plugin
 {
+
     // Static Properties
-    // =========================================================================
 
     /**
-     * Static property that is an instance of this plugin class so that it can be accessed via
-     * BlurHash::$plugin
-     *
      * @var BlurHash
      */
     public static $plugin;
 
     // Public Properties
-    // =========================================================================
 
     /**
-     * To execute your plugin’s migrations, you’ll need to increase its schema version.
-     *
      * @var string
      */
     public $schemaVersion = '1.0.0';
 
     /**
-     * Set to `true` if the plugin should have a settings view in the control panel.
-     *
      * @var bool
      */
     public $hasCpSettings = false;
 
     /**
-     * Set to `true` if the plugin should have its own section (main nav item) in the control panel.
-     *
      * @var bool
      */
     public $hasCpSection = false;
 
     // Public Methods
-    // =========================================================================
 
-    /**
-     * Set our $plugin static property to this class so that it can be accessed via
-     * BlurHash::$plugin
-     *
-     * Called after the plugin class is instantiated; do any one-time initialization
-     * here such as hooks and events.
-     *
-     * If you have a '/vendor/autoload.php' file, it will be loaded for you automatically;
-     * you do not need to load it in your init() method.
-     *
-     */
     public function init()
     {
         parent::init();
         self::$plugin = $this;
 
+        $this->setComponents([
+            'blurHashServices' => BlurHashService::class,
+        ]);
+        
         // Add in our Twig extensions
         Craft::$app->view->registerTwigExtension(new BlurHashTwigExtension());
 
-        // Do something after we're installed
         Event::on(
-            Plugins::class,
-            Plugins::EVENT_AFTER_INSTALL_PLUGIN,
-            function (PluginEvent $event) {
-                if ($event->plugin === $this) {
-                    // We were just installed
-                }
+            Gql::class,
+            Gql::EVENT_REGISTER_GQL_DIRECTIVES,
+            function(RegisterGqlDirectivesEvent $event) {
+                $event->directives[] = BlurHashToUri::class;
+                $event->directives[] = AssetToBlurHash::class;
             }
         );
 
-/**
- * Logging in Craft involves using one of the following methods:
- *
- * Craft::trace(): record a message to trace how a piece of code runs. This is mainly for development use.
- * Craft::info(): record a message that conveys some useful information.
- * Craft::warning(): record a warning message that indicates something unexpected has happened.
- * Craft::error(): record a fatal error that should be investigated as soon as possible.
- *
- * Unless `devMode` is on, only Craft::warning() & Craft::error() will log to `craft/storage/logs/web.log`
- *
- * It's recommended that you pass in the magic constant `__METHOD__` as the second parameter, which sets
- * the category to the method (prefixed with the fully qualified class name) where the constant appears.
- *
- * To enable the Yii debug toolbar, go to your user account in the AdminCP and check the
- * [] Show the debug toolbar on the front end & [] Show the debug toolbar on the Control Panel
- *
- * http://www.yiiframework.com/doc-2.0/guide-runtime-logging.html
- */
         Craft::info(
             Craft::t(
                 'blur-hash',
@@ -132,7 +91,11 @@ class BlurHash extends Plugin
         );
     }
 
-    // Protected Methods
-    // =========================================================================
+    // Settings
+
+    protected function createSettingsModel()
+    {
+        return new Settings();
+    }
 
 }