+++ src/translations/en/blur-hash.php
@@ -11,11 +11,6 @@
 /**
  * BlurHash en Translation
  *
- * Returns an array with the string to be translated (as passed to `Craft::t('blur-hash', '...')`) as
- * the key, and the translation as the value.
- *
- * http://www.yiiframework.com/doc-2.0/guide-tutorial-i18n.html
- *
  * @author    Dodeca Studio
  * @package   BlurHash
  * @since     1.0.0